wip: upgrade kava to sdk v0.50
